Fix document viewer crashes and project page failures (closes #61)
- Fix RLS infinite recursion (42P17): replace self-referential
  project_members SELECT policy with get_user_project_role() SECURITY
  DEFINER call to break the recursion loop
- Fix compliance API: always return issues:[] so ComplianceReport
  never calls .filter() on undefined
- Fix DOMPurify SSR crash: guard sanitize() call with
  typeof window !== 'undefined' check; move helper after imports
- Fix i18n: replace {{count}} with {count} (ICU format) in en/no messages
- Fix middleware: skip intlMiddleware for /api/ routes to prevent
  locale prefix being prepended to API calls
- Refactor ProjectPage to async server component; extract interactive
  parts into ProjectClient to fix browser Supabase auth session issue
